src/main/java/hello/itemservice/domain/Member/MemberMemoryRepository.java
package hello.itemservice.domain.Member;

import org.springframework.stereotype.Repository;

import java.util.ArrayList;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import java.util.concurrent.ConcurrentHashMap;

@Repository
public class MemberMemoryRepository implements MemberRepository {
    private static Map<Long, Member> store = new ConcurrentHashMap<>();
    private static Long sequence = 0L;

    @Override
    public Member save(Member member) {
        member.setMemberId(++sequence);
        store.put(member.getMemberId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long memberId) {
        return Optional.ofNullable(store.get((memberId)));
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    @Override
    public Optional<Member> findByName(String memberName) {
        return store.values().stream()
                .filter(findMember -> findMember.getMemberName().equals(memberName))
                .findAny();
    }

    public void clearStore() {
        store.clear();
    }
}
